解决判空问题

+++ mall/src/main/java/com/cn/wanxi/mall/controller/menu/MenuController.java
@@ -1,16 +1,21 @@
 package com.cn.wanxi.mall.controller.menu;
 
+import com.cn.wanxi.entity.brand.BrandEntity;
 import com.cn.wanxi.entity.brand.ByPage;
 import com.cn.wanxi.entity.brand.PageList;
 import com.cn.wanxi.entity.menu.MenuEntity;
 import com.cn.wanxi.service.menu.IMenuService;
 import com.cn.wanxi.utils.utils.Msg;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.util.ObjectUtils;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 import java.util.Map;
+
+import static org.springframework.util.ObjectUtils.isEmpty;
 
 /**
  * 【菜单管理】 菜单为三级菜单
@@ -73,11 +78,14 @@ public class MenuController {
     public Msg findById(@RequestBody Map<String, Integer> param) {
         Msg msg = null;
         int id = param.get("id");
-        MenuEntity byId = iMenuService.findById(id);
-        if (byId != null) {
-            msg = Msg.success().messageData(byId);
-        } else {
-            msg = Msg.fail();
+        if (!StringUtils.isEmpty(id) && id > 0) {
+            MenuEntity byId = iMenuService.findById(id);
+            //判断是否有返回的数据
+            if (!ObjectUtils.isEmpty(byId)) {
+                msg = Msg.success().messageData(byId);
+            } else {
+                msg = Msg.fail().messageData("该品牌不存在");
+            }
         }
         return msg;
     }
@@ -95,7 +103,7 @@ public class MenuController {
         if (byName != null) {
             msg = Msg.success().messageData(byName);
         } else {
-            msg = Msg.fail();
+            msg = Msg.fail().messageData("请输入正确的名字");
         }
         return msg;
     }
@@ -106,13 +114,23 @@ public class MenuController {
      * @return
      */
     @RequestMapping(value = "/update", method = RequestMethod.POST, produces = "application/json;charset=UTF-8")
-    public Msg updateInfo(@RequestBody MenuEntity menuEntity) {
+    public Msg update(@RequestBody MenuEntity menuEntity) {
         Msg msg = null;
-        int up = iMenuService.update(menuEntity);
-        if (up > 0) {
-            msg = Msg.success().messageData(menuEntity);
+        int id = menuEntity.getId();
+        if (id > 0) {
+            //根据id查询数据
+            MenuEntity byId = iMenuService.findById(id);
+            //判断是否查询到该品牌信息
+            if (!ObjectUtils.isEmpty(byId)) {
+                int result = iMenuService.update(menuEntity);
+                if (result > 0) {
+                    msg = Msg.success().messageData(menuEntity);
+                }
+            } else {
+                msg = Msg.fail().messageData("该品牌不存在");
+            }
         } else {
-            msg = Msg.fail();
+            msg = Msg.fail().messageData("请输入id");
         }
         return msg;
     }
@@ -127,11 +145,15 @@ public class MenuController {
         response.setHeader("Access-Control-Allow-Origin", "*");
         Msg msg = null;
         int id = param.get("id");
-        int i = iMenuService.deleteById(id);
-        if (i > 0) {
-            msg = Msg.success();
+        if (id > 0) {
+            int i = iMenuService.deleteById(id);
+            if (i > 0) {
+                msg = Msg.success().messageData("删除成功");
+            } else {
+                msg = Msg.fail().messageData("删除失败,该用户不存在");
+            }
         } else {
-            msg = Msg.fail();
+            msg = Msg.fail().messageData("请输入id");
         }
         return msg;
     }
@@ -157,7 +179,7 @@ public class MenuController {
 //        }
 
         //实例化 分页实体类
-        PageList pageList = new PageList();
+        com.cn.wanxi.entity.brand.PageList pageList = new PageList();
         //根据页数，每页记录数查询
         List<Map<String, Object>> list = iMenuService.findListAndPage(menuEntity, page, size);
         //把查询出来的对象封装在分页实体类中
